app/ping-web/requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Django==4.2.2
django-admin-sortable2==2.1.10
django-js-asset==2.1.0
django-mdeditor==0.1.20
django-mptt==0.15.0
django-redis==5.3.0
exceptiongroup==1.1.3
frozenlist==1.4.0
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
line-bot-sdk==3.1.0
Markdown==3.5
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.13
PySocks==1.7.1
python-dateutil==2.8.2
redis==5.0.0
requests==2.31.0
selenium==4.15.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
